+++ src/cpath.c
@@ -9,15 +9,15 @@ void cpath_init(struct cpath* path, struct dp_state_table const* state_tbl, stru
     path->bits_trans = malloc(sizeof(*path->bits_trans) * (unsigned)path->nstates);
     path->bits_state[0] = 0;
 
-    for (uint_fast16_t tgt = 0; tgt < path->nstates; ++tgt) {
+    for (unsigned tgt = 0; tgt < path->nstates; ++tgt) {
 
-        uint_fast8_t depth = 0;
-        for (uint_fast16_t i = 0; i < dp_trans_table_ntrans(trans_tbl, tgt); ++i) {
+        unsigned depth = 0;
+        for (unsigned i = 0; i < dp_trans_table_ntrans(trans_tbl, tgt); ++i) {
 
-            uint_fast16_t src = dp_trans_table_source_state(trans_tbl, tgt, i);
-            uint_fast8_t  min_seq = dp_state_table_min_seq(state_tbl, src);
-            uint_fast8_t  max_seq = dp_state_table_max_seq(state_tbl, src);
-            depth = (uint_fast8_t)MAX(max_seq - min_seq, depth);
+            unsigned src = dp_trans_table_source_state(trans_tbl, tgt, i);
+            unsigned min_seq = dp_state_table_min_seq(state_tbl, src);
+            unsigned max_seq = dp_state_table_max_seq(state_tbl, src);
+            depth = MAX(max_seq - min_seq, depth);
         }
         path->bits_state[tgt + 1] = path->bits_state[tgt];
         path->bits_trans[tgt] = (uint8_t)bits_width((uint32_t)dp_trans_table_ntrans(trans_tbl, tgt));
