+++ src/dp_matrix.c
@@ -16,12 +16,12 @@ struct dp_matrix* dp_matrix_create(struct dp_state_table const* states)
         return NULL;
     }
 
-    uint_fast16_t next_col = 0;
-    for (uint_fast16_t i = 0; i < dp_state_table_nstates(states); ++i) {
-        uint_fast8_t const min = dp_state_table_min_seq(states, i);
-        uint_fast8_t const max = dp_state_table_max_seq(states, i);
+    unsigned next_col = 0;
+    for (unsigned i = 0; i < dp_state_table_nstates(states); ++i) {
+        unsigned const min = dp_state_table_min_seq(states, i);
+        unsigned const max = dp_state_table_max_seq(states, i);
         matrix->state_col[i] = (int16_t)(next_col - min);
-        next_col += (uint_fast16_t)(max - min + 1);
+        next_col += (unsigned)(max - min + 1);
     }
 
     matrix->score = matrixf_create(MAX_LOOKUP, next_col);
