+++ src/dp_matrix.h
@@ -23,20 +23,20 @@ struct dp_matrix
 
 struct dp_matrix*       dp_matrix_create(struct dp_state_table const* states);
 void                    dp_matrix_destroy(struct dp_matrix const* matrix);
-static inline imm_float dp_matrix_get_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
-                                            uint_fast8_t seq_len);
-static inline void      dp_matrix_set_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
-                                            uint_fast8_t seq_len, imm_float score);
+static inline imm_float dp_matrix_get_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, unsigned state,
+                                            unsigned seq_len);
+static inline void      dp_matrix_set_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, unsigned state,
+                                            unsigned seq_len, imm_float score);
 
-static inline imm_float dp_matrix_get_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
-                                            uint_fast8_t seq_len)
+static inline imm_float dp_matrix_get_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, unsigned state,
+                                            unsigned seq_len)
 {
     uint_fast32_t col = (uint_fast32_t)(dp_matrix->state_col[state] + (int_fast32_t)seq_len);
     return matrixf_get(dp_matrix->score, row % matrixf_nrows(dp_matrix->score), col);
 }
 
-static inline void dp_matrix_set_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
-                                       uint_fast8_t seq_len, imm_float score)
+static inline void dp_matrix_set_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, unsigned state,
+                                       unsigned seq_len, imm_float score)
 {
     uint_fast32_t col = (uint_fast32_t)(dp_matrix->state_col[state] + (int_fast32_t)seq_len);
     matrixf_set(dp_matrix->score, row % matrixf_nrows(dp_matrix->score), col, score);
